Add: Separate TOC section height
When the length of list over view-height,
the TOC scroll rely on the main content height.

Thus, if I want to see the last article of
the TOC, I have to scroll down up to bottom of
the main content, therefore, I have separated
the TOC section from the main content for
acting independently.